requirements.txt
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
